requirements.txt
beautifulsoup4==4.9.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
decorator==4.4.2
future==0.18.2
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.9
Pillow==7.1.2
pycparser==2.20
python-telegram-bot==12.7
qrcode==6.1
requests==2.23.0
six==1.14.0
soupsieve==2.0
tornado==6.0.4
urllib3==1.25.9
